test: clean-up tests
