+++ package.json
@@ -113,6 +113,7 @@
     "@types/node": "^16.18.126",
     "@types/sha.js": "^2.4.4",
     "@types/sinon": "^10.0.20",
+    "@types/sinon-chai": "^4.0.0",
     "@types/source-map-support": "^0.5.10",
     "@types/yargs": "^17.0.33",
     "better-sqlite3": "^8.7.0",
