--- test/functional/connection-options-reader/configs/ormconfig.env
@@ -1,2 +1,0 @@
-TYPEORM_CONNECTION = mysql
-TYPEORM_DATABASE = test-ormconfig-env
