+++ test/github-issues/10626/issue-10626.ts
@@ -1,10 +1,10 @@
-import "reflect-metadata"
-import {
-    createTestingConnections,
-    closeTestingConnections,
-} from "../../utils/test-utils"
-import { DataSource } from "../../../src/index.js"
 import { expect } from "chai"
+import "reflect-metadata"
+import { DataSource } from "../../../src/index.js"
+import {
+    closeTestingConnections,
+    createTestingConnections,
+} from "../../utils/test-utils"
 
 describe("github issues > #10626 Postgres CREATE INDEX CONCURRENTLY bug", () => {
     let dataSources: DataSource[]
@@ -16,7 +16,6 @@ describe("github issues > #10626 Postgres CREATE INDEX CONCURRENTLY bug", () => 
                 schemaCreate: false,
                 dropSchema: true,
                 enabledDrivers: ["postgres"],
-                logging: true,
             })),
     )
