+++ test/unit/connection-options-reader/configs/test-path-config-async.ts
