+++ test/utils/test-setup.ts
@@ -1,10 +1,13 @@
 import "source-map-support/register"
 import "reflect-metadata"
+
 import chai from "chai"
+import sinonChai from "sinon-chai"
+import chaiAsPromised from "chai-as-promised"
 
 // Tests assume UTC time zone when formatting/parsing dates.
 process.env.TZ = "UTC"
 
 chai.should()
-chai.use(require("sinon-chai"))
-chai.use(require("chai-as-promised"))
+chai.use(sinonChai)
+chai.use(chaiAsPromised)
